Add Week 1 Day 2 task for Babalolas Todo Task

+++ submissions/week-1/day-5/Babalola-Taiwo_Day5-Tasks1-2/Task-2/src/main.rs
@@ -26,9 +26,9 @@ fn main() {
 
 
 enum Book {
-    Fiction { title: String, author: String, price: f32 },
-    Magazine { title: String, author: String, price: f32 },
-    SciFi { title: String, price: f32 },
+    Fiction { title: String, author: String, price: f64 },
+    Magazine { title: String, author: String, price: f64 },
+    SciFi { title: String, price: f64 },
 }
 
 fn print_book_details(books: Vec<Book>) {
